Improve recorded detail of exception stacktraces (#70)
In many X-Ray SDK use cases, including the Django/Flask middleware and
aws_xray_sdk.core.xray_recorder.capture and .capture_async cases, there
is a broad try/except "catch-all" for exceptions which records useful
information in an X-Ray trace in the event of an unexpected exception.

However, the approach (prior to this patch) of using
`traceback.extract_stack` in the `except` block to extract stack trace
information isn't terribly useful when viewing X-Ray traces for
debugging purposes. `traceback.extract_stack` only shows the stack state
up to and including the `traceback.extract_stack` call in the `except`
block; it does not include the stack trace information from the caught
exception--which are arguably the most important details to include in
the X-Ray trace! Similarly, including _only_ the exception's stack trace
information can omit important code path context if the
exeception-producing code is wrapped in a decorator. This is a very
subtle but important detail. There is a good discussion of exactly this
problem--specificaly in the context of decorators with a "catch-all"
try/except--on Stackoverflow:
https://stackoverflow.com/questions/14527819/traceback-shows-up-until-decorator.

This patch introduces the `stacktrace` utility module and is an attempt
to include more relevant info in recorded X-Ray traces while respecting
the `max_trace_back` behavior expressed in many places to avoid bloated
trace messages.

To illustrate the differencethis makes, consider a trivial Lambda
function written in Python with a single module called `myfunction.py`:

```
from aws_xray_sdk.core import xray_recorder

@xray_recorder.capture('handler')
def handler(event, context):
    data = foo()
    return data

def foo():
    return bar()

def bar():
    raise Exception('Something went wrong!')
    return {'data': [1,2,3]}
```

There are multiple function calls here in order to create an
interesting call stack for tracing purposes.

With the current implementation of aws_xray_sdk, here's the trace that
gets recorded by X-Ray:

    Exception: Something went wrong!
    at <module> (bootstrap.py:538)
    at main (bootstrap.py:533)
    at handle_event_request (bootstrap.py:250)

With this patch applied, we get a lot more information about the call
stack, as we would expect:

    Exception: Something went wrong!
    at <module> (bootstrap.py:538)
    at main (bootstrap.py:533)
    at handle_event_request (bootstrap.py:250)
    at handler (myfunction.py:6)
    at foo (myfunction.py:11)
    at bar (myfunction.py:15)

With this we can much more easily trace the origin of a runtime error.