Remove the dependency on requests
This removes the use of `requests` from the core functionality, which
results in `requests` being only used as extension, therefore not being
necessary as dependency anymore.

The motivation for this change is to get rid of all the dependencies of
`requests`, whose size sums up quite a bit, when not requiring to patch
`requests`. It follows the removal of `requests` from `botocore`, which
was done as part of v1.11.0.